refactor: move service code outside of main code
